refactor(core): group io helpers
Move cache and remote download helpers into a core io folder so the core root keeps only orchestration and small option/runtime utilities.
