+++ src/core/io/cache.ts
